pyproject.toml
[tool.poetry]
name = "cr-apk"
packages = [{ include = "src" }]
version = "v0.0.0"
description = "APK Manager for Clash Royale"
readme = "README.md"
authors = ["Martin Miglio <code@martinmiglio.dev>"]
repository = "https://github.com/pyclashbot/cr-apk"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
cleo = "^2.1.0"
python-dotenv = "^1.0.1"
pymemuc = "^0.6.0"
boto3 = "^1.35.37"
boto3-stubs = {extras = ["s3"], version = "^1.35.37"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[tool.poetry.scripts]
crapk = "src.main:main"

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff]
line-length = 120
